feat(boundary): group boundaries in shared/ui, make SuspenseQueryBoundary
Related to: #63
