+++ src/features/checkAuthorization/ui/RequireAuthorizationWrapper.tsx
@@ -2,7 +2,7 @@ import { ReactNode, Suspense } from "react";
 
 import { getMeApi } from "@/entities/user/api/user.server";
 
-import FetchBoundary from "@/shared/ui/FetchBoundary";
+import FetchBoundary from "@/shared/ui/Boundary/FetchBoundary";
 
 import RequireAuthorizationButton from "./RequireAuthorizationButton";
