+++ src/shared/ui/Boundary/FetchBoundary.tsx
