+++ src/views/MyPage/ConnectCouple/ui/index.tsx
@@ -8,7 +8,7 @@ import { withAuthorizationRoute } from "@/features/handleAuthorizationRoute/ui";
 import { getMeApi } from "@/entities/user/api/user.server";
 
 import BackButton from "@/shared/ui/BackButton";
-import FetchBoundary from "@/shared/ui/FetchBoundary";
+import FetchBoundary from "@/shared/ui/Boundary/FetchBoundary";
 import DvhHeightLayout from "@/shared/ui/Layout/DvhHeightLayout";
 
 import { IconArrowLeft24, IconCharacterInvite } from "public/icons";
