+++ src/views/MyPage/UnconnectedCouple/ui/index.tsx
@@ -5,8 +5,8 @@ import { withAuthorizationRoute } from "@/features/handleAuthorizationRoute/ui";
 import { getMeApi } from "@/entities/user/api/user.server";
 
 import BackButton from "@/shared/ui/BackButton";
+import FetchBoundary from "@/shared/ui/Boundary/FetchBoundary";
 import ChevronLink from "@/shared/ui/ChevronLink";
-import FetchBoundary from "@/shared/ui/FetchBoundary";
 
 import RestoreCoupleLink from "./RestoreCoupleLink";
 import { IconArrowLeft24 } from "public/icons";
